[WIP] Add bolt task timeout parameter
Allows the user to override the bolt task DDL timeout and specify
how long the client application and agent should wait synchronously
before returning with tasks left running in the background.

The usecase for this is running bolt tasks from a CI system, and
having the mco task run command block until the task is done
(provided a suitable timeout is set).

Works so far:
- lowering the timeout below the default of 60s

Not done yet:
- Support for increasing the timeout beyond the default 60s
- Determining whether this should be a separate parameter, or
  whether the mco rpc `timeout` parameter should be passed through
  instead
- Tests
- Docs